refactor: rename the archive
